Revert "check check, is anyone there? *static*"
Revert submission 17016040-change-17016040

Reason for revert: DroidMonitor-triggered revert due to breakage , bug b/223737148
BUG: 223737148
Reverted Changes:
I6cdcee15a:Fix bluetooth and aac dependency
I0d232c8bc:check check, is anyone there? *static*

Test: N/A, this is a rollback

+++ system/gd/common/byte_array_test.cc
@@ -31,7 +31,7 @@ static uint8_t data[16] = {
 TEST(ByteArrayTest, test_constructor_array) {
   ByteArray<16> byte_array(data);
 
-  for (size_t i = 0; i < ByteArray<16>::kLength; i++) {
+  for (int i = 0; i < ByteArray<16>::kLength; i++) {
     ASSERT_EQ(data[i], byte_array.bytes[i]);
   }
 }
@@ -40,7 +40,7 @@ TEST(ByteArrayTest, test_from_str) {
   auto byte_array = ByteArray<16>::FromString(test_bytes);
   ASSERT_TRUE(byte_array);
 
-  for (size_t i = 0; i < ByteArray<16>::kLength; i++) {
+  for (int i = 0; i < ByteArray<16>::kLength; i++) {
     ASSERT_EQ(test_data[i], byte_array->bytes[i]);
   }
 }
